+++ android/app/src/main/java/space/holoboard/app/DelegationService.java
@@ -1,0 +1,13 @@
+package space.holoboard.app;
+
+
+
+public class DelegationService extends
+        com.google.androidbrowserhelper.trusted.DelegationService {
+    @Override
+    public void onCreate() {
+        super.onCreate();
+
+
+    }
+}